Improved groups and accounts callback, added buttons

+++ tg-bot/src/main/java/polis/bot/Bot.java
@@ -70,12 +70,14 @@ import java.util.Map;
 import java.util.Objects;
 import java.util.stream.Collectors;
 
+import static polis.commands.NonCommand.VK_GROUP_ADDED;
 import static polis.datacheck.OkDataCheck.OK_AUTH_STATE_ANSWER;
 import static polis.datacheck.OkDataCheck.OK_AUTH_STATE_SERVER_EXCEPTION_ANSWER;
 import static polis.datacheck.OkDataCheck.OK_AUTH_STATE_WRONG_AUTH_CODE_ANSWER;
 import static polis.datacheck.OkDataCheck.OK_GROUP_ADDED;
 import static polis.datacheck.OkDataCheck.USER_HAS_NO_RIGHTS;
 import static polis.datacheck.OkDataCheck.WRONG_LINK_OR_USER_HAS_NO_RIGHTS;
+import static polis.datacheck.VkDataCheck.VK_AUTH_STATE_ANSWER;
 import static polis.keyboards.Keyboard.GO_BACK_BUTTON_TEXT;
 import static polis.telegram.TelegramDataCheck.BOT_NOT_ADMIN;
 import static polis.telegram.TelegramDataCheck.RIGHT_LINK;
@@ -89,28 +91,24 @@ public class Bot extends TelegramLongPollingCommandBot implements TgFileLoader, 
             + "курсе автоматически опубликованных записей с помощью команды /notifications";
     private static final String AUTOPOSTING_ENABLE_AND_NOTIFICATIONS = "Функция автопостинга включена."
             + TURN_ON_NOTIFICATIONS_MSG;
-    private static final Map<String, List<String>> BUTTONS_TEXT_MAP = Map.of(
-            String.format(OK_AUTH_STATE_ANSWER, State.OkAccountDescription.getIdentifier()),
-            List.of(State.OkAccountDescription.getDescription()),
-            String.format(OK_GROUP_ADDED, State.SyncOkTg.getIdentifier()),
-            List.of(State.SyncOkTg.getDescription()),
-            RIGHT_LINK,
-            List.of(State.TgChannelDescription.getDescription()),
-            OK_AUTH_STATE_WRONG_AUTH_CODE_ANSWER,
-            EMPTY_LIST,
-            OK_AUTH_STATE_SERVER_EXCEPTION_ANSWER,
-            EMPTY_LIST,
-            WRONG_LINK_OR_USER_HAS_NO_RIGHTS,
-            EMPTY_LIST,
-            USER_HAS_NO_RIGHTS,
-            EMPTY_LIST,
-            WRONG_LINK_OR_BOT_NOT_ADMIN,
-            EMPTY_LIST,
-            BOT_NOT_ADMIN,
-            EMPTY_LIST,
-            AUTOPOSTING_ENABLE_AND_NOTIFICATIONS,
-            List.of(State.Notifications.getDescription())
-    );
+    private static final Map<String, List<String>> BUTTONS_TEXT_MAP = Map.ofEntries(
+            Map.entry(String.format(OK_AUTH_STATE_ANSWER, State.OkAccountDescription.getIdentifier()),
+                    List.of(State.OkAccountDescription.getDescription())),
+            Map.entry(String.format(OK_GROUP_ADDED, State.SyncOkTg.getIdentifier()),
+                    List.of(State.SyncOkTg.getDescription())),
+            Map.entry(RIGHT_LINK, List.of(State.TgChannelDescription.getDescription())),
+            Map.entry(OK_AUTH_STATE_WRONG_AUTH_CODE_ANSWER, EMPTY_LIST),
+            Map.entry(OK_AUTH_STATE_SERVER_EXCEPTION_ANSWER, EMPTY_LIST),
+            Map.entry(WRONG_LINK_OR_USER_HAS_NO_RIGHTS, EMPTY_LIST),
+            Map.entry(USER_HAS_NO_RIGHTS, EMPTY_LIST),
+            Map.entry(WRONG_LINK_OR_BOT_NOT_ADMIN, EMPTY_LIST),
+            Map.entry(BOT_NOT_ADMIN, EMPTY_LIST),
+            Map.entry(AUTOPOSTING_ENABLE_AND_NOTIFICATIONS, List.of(State.Notifications.getDescription())),
+            Map.entry(String.format(VK_AUTH_STATE_ANSWER, State.VkAccountDescription.getIdentifier()),
+                    List.of(State.VkAccountDescription.getDescription())),
+            Map.entry(String.format(VK_GROUP_ADDED, State.SyncVkTg.getIdentifier()),
+                    List.of(State.SyncVkTg.getDescription())
+    ));
     private final String botName;
     private final String botToken;
     private static final Logger LOGGER = LoggerFactory.getLogger(Bot.class);
@@ -508,13 +506,16 @@ public class Bot extends TelegramLongPollingCommandBot implements TgFileLoader, 
                 }
             }
             case GROUP_CALLBACK_TEXT -> {
+                if (dataParts.length < 4) {
+                    LOGGER.error(String.format("Wrong group-callback data: %s", data));
+                    return;
+                }
                 if (Objects.equals(dataParts[2], "0")) {
                     changeCurrentSocialMediaGroupAndExecuteCommand(chatId, dataParts, msg, State.GroupDescription);
                 } else if (Objects.equals(dataParts[2], "1")) {
                     CurrentChannel currentChannel = currentChannelRepository.getCurrentChannel(chatId);
-                    CurrentAccount currentAccount = currentAccountRepository.getCurrentAccount(chatId);
-                    channelGroupsRepository.deleteChannelGroup(currentChannel.getChannelId(),
-                            currentAccount.getSocialMedia());
+                    String socialMediaName = dataParts[2];
+                    channelGroupsRepository.deleteChannelGroup(currentChannel.getChannelId(), socialMediaName);
                     currentGroupRepository.deleteCurrentGroup(chatId);
                     deleteLastMessage(msg, chatId);
                     getRegisteredCommand(State.TgSyncGroups.getIdentifier()).processMessage(this, msg, null);
@@ -523,14 +524,14 @@ public class Bot extends TelegramLongPollingCommandBot implements TgFileLoader, 
                 }
             }
             case ACCOUNT_CALLBACK_TEXT -> {
-                if (dataParts.length < 3) {
+                if (dataParts.length < 4) {
                     LOGGER.error(String.format("Wrong account-callback data: %s", data));
                     return;
                 }
                 boolean shouldDelete = dataParts[2].equals("1");
-                String currentAccountSocialMedia = currentAccountRepository.getCurrentAccount(chatId).getSocialMedia();
+                String socialMediaName = dataParts[3];
                 State state = shouldDelete ? State.AddGroup :
-                        (currentAccountSocialMedia.equals(SocialMedia.OK.getName()) ? State.OkAccountDescription
+                        (socialMediaName.equals(SocialMedia.OK.getName()) ? State.OkAccountDescription
                                 : State.VkAccountDescription);
                 processAccountCallback(msg, chatId, dataParts, state, shouldDelete);
                 currentStateRepository.insertCurrentState(new CurrentState(
